refactor: DetailResponse 내부에 관련 record 를 선언한다.
