--- backend/src/main/java/com/festago/festival/dto/DetailTicketResponse.java
@@ -1,9 +1,0 @@
-package com.festago.festival.dto;
-
-public record DetailTicketResponse(
-    Long id,
-    String ticketType,
-    Integer totalAmount,
-    Integer remainAmount) {
-
-}
