+++ backend/src/main/java/com/festago/stage/application/FestivalStageServiceImpl.java
@@ -9,8 +9,8 @@ import com.festago.common.exception.NotFoundException;
 import com.festago.festival.application.FestivalStageService;
 import com.festago.festival.domain.Festival;
 import com.festago.festival.dto.DetailFestivalResponse;
-import com.festago.festival.dto.DetailStageResponse;
-import com.festago.festival.dto.DetailTicketResponse;
+import com.festago.festival.dto.DetailFestivalResponse.DetailStageResponse;
+import com.festago.festival.dto.DetailFestivalResponse.DetailStageResponse.DetailTicketResponse;
 import com.festago.festival.repository.FestivalRepository;
 import com.festago.stage.domain.Stage;
 import com.festago.stage.repository.StageRepository;
